% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctll.R
\name{ctll}
\alias{ctll}
\title{Continuous time Local Level model}
\usage{
ctll(y, u = NULL, type = c("stock", "flow"), log = TRUE, h = 12,
  silent = TRUE, B = c(0.1, 0.1))
}
\arguments{
\item{y}{Vector or ts object, containing data needed to be forecasted.}

\item{type}{What type of variable to assume: stock or flow. Example of the stock
variable is a price of a product. Example of the flow variable is the income over
time.}

\item{log}{Whether to take logarithms of the demand sizes or not.}

\item{h}{Length of forecasting horizon.}

\item{silent}{Specifies, whether to provide the progress of the function
or not. If TRUE, then the function will print what it does and how much it
has already done.}

\item{B}{The vector of initial parameters (variances).}
}
\description{
Function applies and estimates the continuous time local level model.
}
\details{
Function applies a discretised version of the continuous time local level model.
This model can be applied to the demand that happens at irregular frequency. It
considers the zero values in the data as the absence of the value and only focuses
on the demand sizes of the data.
}
\examples{
y <- rpois(100,1)
ctll(y)

}
\seealso{
\code{\link[UComp]{UC}}, \code{\link[smooth]{adam}}
}
\author{
Diego J. Pedregal, \email{Diego.Pedregal@uclm.es}

Ivan Svetunkov, \email{ivan@svetunkov.com}
}
\keyword{models}
\keyword{nonlinear}
\keyword{regression}
\keyword{smooth}
\keyword{ts}
\keyword{univar}
