% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTSS3functions.R
\name{fitted.PTS}
\alias{fitted.PTS}
\title{fitted.PTS}
\usage{
\method{fitted}{PTS}(object, ...)
}
\arguments{
\item{object}{Object of class \dQuote{PTS}.}

\item{...}{Additional inputs to function.}
}
\description{
Fitted output values of PTS object
}
\details{
See help of \code{PTS}.
}
\examples{
m1 <- PTSmodel(log(AirPassengers))
fitted(m1)
}
\seealso{
\code{\link{PTS}}, \code{\link{PTSmodel}}, \code{\link{PTSvalidate}},
         \code{\link{PTScomponents}}, \code{\link{PTSestim}}
}
\author{
Diego J. Pedregal
}
