% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTSfunctions.R
\name{modelUC2PTS}
\alias{modelUC2PTS}
\title{modelUC2PTS}
\usage{
modelUC2PTS(modelUC, lambda)
}
\arguments{
\item{modelUC}{a UC model}

\item{lambda}{is the Box-Cox parameter}
}
\value{
a PTS model
}
\description{
Translates modelUC to model PTS
}
\author{
Diego J. Pedregal, \email{Diego.Pedregal@uclm.es}

Ivan Svetunkov, \email{ivan@svetunkov.com}
}
