% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSOEfunctions.R
\name{MSOE}
\alias{MSOE}
\title{MSOE}
\usage{
MSOE(sys)
}
\value{
An object of class \code{MSOE}. It is a list with fields including all the inputs and
        the fields listed below as outputs. All the functions in this package fill in
        part of the fields of any \code{MSOE} object as specified in what follows (function
        \code{MSOE} fills in all of them at once):

After running \code{MSOEmodel} or \code{MSOEestim}:
\itemize{
\item p:        Estimated parameters
\item v:        Estimated innovations (white noise in correctly specified models)
\item yFor:     Forecasted values of output
\item yForV:    Forecasted values +- one standard error
\item criteria: Value of criteria for estimated model
\item iter:     Number of iterations in estimation
\item grad:     Gradient at estimated parameters
\item covp:     Covariance matrix of parameters
}

After running \code{MSOEvalidate}:
\itemize{
\item table: Estimation and validation table
}

After running \code{MSOEcomponents}:
\itemize{
\item comp:  Estimated components in matrix form
\item compV: Estimated components variance in matrix form
}

After running \code{MSOEfilter}, \code{MSOEsmooth} or  \code{MSOEdisturb}:
\itemize{
\item yFit:  Fitted values of output
\item yFitV: Variance of fitted values of output
\item a:     State estimates
\item P:     Variance of state estimates
\item aFor:  Forecasts of states
\item PFor:  Forecasts of states variances
}

After running \code{MSOEdisturb}:
\itemize{
\item eta: State perturbations estimates
\item eps: Observed perturbations estimates
}
}
\description{
Runs all relevant functions for MSOE modelling
}
\details{
\code{MSOE} is a function for modelling and forecasting univariate
time series according to Unobserved Components models (MSOE).
It sets up the model with a number of control variables that
govern the way the rest of functions in the package work. It also estimates
the model parameters by Maximum Likelihood, forecasts the data, performs smoothing,
estimates model disturbances, estimates components and shows statistical diagnostics.
Standard methods applicable to MSOE objects are print, summary, plot,
fitted, residuals, logLik, AIC, BIC, coef, predict, tsdiag.
}
\examples{
\dontrun{
y <- log(AirPassengers)
m1 <- MSOE(y)
m1 <- MSOE(y, model = "llt/different/arma(0,0)")
}
}
\seealso{
\code{\link{MSOE}}, \code{\link{MSOEvalidate}}, \code{\link{MSOEfilter}}, \code{\link{MSOEsmooth}},
         \code{\link{MSOEcomponents}},
}
\author{
Diego J. Pedregal, \email{Diego.Pedregal@uclm.es}

Ivan Svetunkov, \email{ivan@svetunkov.com}
}
