% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSOEfunctions.R
\name{MSOEcomponents}
\alias{MSOEcomponents}
\title{MSOEcomponents}
\usage{
MSOEcomponents(sys)
}
\arguments{
\item{sys}{an object of type \code{MSOE} created with \code{MSOE} or \code{MSOEmodel}}
}
\value{
The same input object with the appropriate fields 
filled in, in particular:
\itemize{
\item comp:  Estimated components in matrix form
\item compV: Estimated components variance in matrix form
}
}
\description{
Estimates unobserved components of MSOE models
Standard methods applicable to MSOEomp objects are print, summary, plot,
fitted, residuals, logLik, AIC, BIC, coef, predict, tsdiag.
}
\examples{
\dontrun{
m1 <- MSOE(log(AirPassengers))
m1 <- MSOEcomponents(m1)
}
}
\seealso{
\code{\link{MSOE}}, \code{\link{MSOEmodel}}, \code{\link{MSOEvalidate}}, \code{\link{MSOEfilter}}, 
         \code{\link{MSOEsmooth}}, \code{\link{MSOEdisturb}},
}
