% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSOEfunctions.R
\name{MSOEestim}
\alias{MSOEestim}
\title{MSOEestim}
\usage{
MSOEestim(sys)
}
\arguments{
\item{sys}{an object of type \code{MSOE} created with \code{MSOE}}
}
\value{
The same input object with the appropriate fields
filled in, in particular:
\itemize{
\item p:        Estimated transformed parameters
\item v:        Estimated innovations (white noise in correctly specified models)
\item yFor:     Forecast values of output
\item yForV:    Forecasted values variance
\item criteria: Value of criteria for estimated model
\item covp:     Covariance matrix of estimated transformed parameters
\item grad:     Gradient of log-likelihood at the optimum
\item iter:     Estimation iterations
}
}
\description{
Estimates and forecasts MSOE models
}
\details{
\code{MSOEestim} estimates and forecasts a time series using an
MSOE model.
The optimization method is a BFGS quasi-Newton algorithm with a
backtracking line search using Armijo conditions.
Parameter names in output table are the following:
\itemize{
\item Damping:   Damping factor for DT trend.
\item Level:     Variance of level disturbance.
\item Slope:     Variance of slope disturbance.
\item Rho(#):    Damping factor of cycle #.
\item Period(#): Estimated period of cycle #.
\item Var(#):    Variance of cycle #.
\item Seas(#):   Seasonal harmonic with period #.
\item Irregular: Variance of irregular component.
\item AR(#):     AR parameter of lag #.
\item MA(#):     MA parameter of lag #.
\item AO#:       Additive outlier in observation #.
\item LS#:       Level shift outlier in observation #.
\item SC#:       Slope change outlier in observation #.
\item Beta(#):   Beta parameter of input #.
\item Cnst:      Constant.
}

Standard methods applicable to MSOEomp objects are print, summary, plot,
fitted, residuals, logLik, AIC, BIC, coef, predict, tsdiag.
}
\examples{
\dontrun{
# m1 <- MSOEsetup(log(AirPassengers))
# m1 <- MSOEestim(m1)
}
}
\seealso{
\code{\link{MSOE}}, \code{\link{MSOEmodel}}, \code{\link{MSOEvalidate}}, \code{\link{MSOEfilter}},
         \code{\link{MSOEsmooth}}, \code{\link{MSOEcomponents}}
}
\author{
Diego J. Pedregal, \email{Diego.Pedregal@uclm.es}

Ivan Svetunkov, \email{ivan@svetunkov.com}
}
