% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSOEfunctions.R
\name{MSOEfilter}
\alias{MSOEfilter}
\title{MSOEfilter}
\usage{
MSOEfilter(sys)
}
\arguments{
\item{sys}{an object of type \code{MSOE} created with \code{MSOE}}
}
\value{
The same input object with the appropriate fields 
filled in, in particular:
\itemize{
\item yFit:  Fitted values of output
\item yFitV: Variance of fitted values of output
\item a:     State estimates
\item P:     Variance of state estimates (diagonal of covariance matrices)
}
}
\description{
Runs the Kalman Filter for MSOE models
Standard methods applicable to \code{MSOE} objects are print, summary, plot,
fitted, residuals, logLik, AIC, BIC, coef, predict, tsdiag.
}
\examples{
\dontrun{
m1 <- MSOE(log(AirPassengers))
m1 <- MSOEfilter(m1)
}
}
\seealso{
\code{\link{MSOE}}, \code{\link{MSOEmodel}}, \code{\link{MSOEvalidate}}, 
         \code{\link{MSOEsmooth}}, \code{\link{MSOEcomponents}},
}
