% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muse-package.R
\name{muse}
\alias{muse}
\alias{muse-package}
\title{muse package}
\description{
Package contains functions implementing Multiple Source of Error state space models for
purposes of time series analysis and forecasting.
}
\details{
\tabular{ll}{ Package: \tab muse\cr Type: \tab Package\cr Date: \tab
2024-10-18 - Inf\cr License: \tab LGPL-2.1 \cr } The following functions are
included in the package:
}
\examples{

x <- rnorm(100,100,10)

}
\seealso{
\code{\link[smooth]{adam}}
}
\author{
Diego J. Pedregal, \email{Diego.Pedregal@uclm.es}

Ivan Svetunkov, \email{ivan@svetunkov.com}
}
\keyword{models}
\keyword{nonlinear}
\keyword{regression}
\keyword{smooth}
\keyword{ts}
\keyword{univar}
