% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTSS3functions.R
\name{plot.PTS}
\alias{plot.PTS}
\title{plot.PTS}
\usage{
\method{plot}{PTS}(x, ...)
}
\arguments{
\item{x}{Object of class \dQuote{PTS}.}

\item{...}{Additional inputs to function.}
}
\description{
Plot components of PTS object
}
\details{
See help of \code{PTS}.
}
\examples{
m1 <- PTS(log(AirPassengers))
plot(m1)
}
\seealso{
\code{\link{PTS}}, \code{\link{PTSmodel}}, \code{\link{PTSvalidate}},
         \code{\link{PTScomponents}}, \code{\link{PTSestim}}
}
\author{
Diego J. Pedregal
}
