% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTSS3functions.R
\name{print.PTS}
\alias{print.PTS}
\title{print.PTS}
\usage{
\method{print}{PTS}(x, ...)
}
\arguments{
\item{x}{Object of class \dQuote{PTS}.}

\item{...}{Additional inputs to handle the way to print output.}
}
\description{
Prints a PTS object
}
\details{
See help of \code{PTS}.
}
\examples{
# m1 <- PTSmodel(log(AirPassengers))
# print(m1)
}
\seealso{
\code{\link{PTS}}, \code{\link{PTSmodel}}, \code{\link{PTSvalidate}},
         \code{\link{PTScomponents}}, \code{\link{PTSestim}}
}
\author{
Diego J. Pedregal
}
